+++ compiler/GHC/StgToJS/Prim.hs
@@ -1191,6 +1191,7 @@ genPrim prof bound ty op = case op of
   VecRemOp _ _ _                    -> unhandledPrimop op
   VecNegOp _ _ _                    -> unhandledPrimop op
   VecAbsOp _ _ _                    -> unhandledPrimop op
+  VecSqrtOp _ _ _                   -> unhandledPrimop op
   VecIndexByteArrayOp _ _ _         -> unhandledPrimop op
   VecReadByteArrayOp _ _ _          -> unhandledPrimop op
   VecWriteByteArrayOp _ _ _         -> unhandledPrimop op